Pattern de vatnumber incluido mas algunas correccionenes

+++ Acme-Hacker-Rank/src/main/java/controllers/AdministratorController.java
@@ -176,7 +176,7 @@ public class AdministratorController extends AbstractController {
 			}
 		} catch (final Exception e) {
 			final Collection<Integer> creditCardsNumbers = this.creditCardService.getAllNumbers();
-			if (creditCardsNumbers.contains(creditcard.getNumber()))
+			if (creditCardsNumbers.contains(creditcard.getNumber()) && creditcard.equals(this.creditCardService.getCreditCardByNumber(creditcard.getNumber())))
 				this.creditCardService.delete(creditcard);
 			result = new ModelAndView("administrator/create");
 			result.addObject("exception", e);
