+++ Acme-Hacker-Rank/src/main/java/controllers/CompanyController.java
@@ -83,7 +83,7 @@ public class CompanyController extends AbstractController {
 			}
 		} catch (final Exception e) {
 			final Collection<Integer> creditCardsNumbers = this.creditCardService.getAllNumbers();
-			if (creditCardsNumbers.contains(creditcard.getNumber()))
+			if (creditCardsNumbers.contains(creditcard.getNumber()) && creditcard.equals(this.creditCardService.getCreditCardByNumber(creditcard.getNumber())))
 				this.creditCardService.delete(creditcard);
 			result = new ModelAndView("company/create");
 			result.addObject("exception", e);
